+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+- `gulp bump` command to increment version in `CHANGELOG.md`, `package.json` and `bower.json`
+
 ## v0.5.0 - 2015-04-15
 ### Added
 - Use `make watch` to first start compile tasks, then watcher task.